feat(registerApp): add test
